create directories for each feature page, separate apis by table
