+++ src/components/Dashboard.js
@@ -17,7 +17,7 @@ import Box from "@material-ui/core/Box";
 import Tooltip from '@material-ui/core/Tooltip';
 import MenuItem from '@material-ui/core/MenuItem';
 import Paper from '@material-ui/core/Paper';
-import NavigationBar from './NavigationBar';
+import NavigationBar from './common/NavigationBar';
 
 const pages = ['Events'];
 const settings = ['Update Profile', 'Logout'];
