+++ src/components/constants/Routes.js
@@ -1,0 +1,1 @@
+// store all page routes here
