+++ src/componentsTest/Events.js
@@ -1,5 +1,5 @@
 import React, { useRef, useState, useEffect } from "react"
-// import { useAuth } from "../contexts/AuthContext"
+// import { useAuth } from "../../contexts/AuthContext"
 import { Link, useHistory } from "react-router-dom"
 
 import Card from '@material-ui/core/Card';
@@ -14,7 +14,7 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import MuiAlert from '@mui/material/Alert';
-// import NavigationBar from './NavigationBar';
+// import NavigationBar from '../common/NavigationBar';
 import Box from "@material-ui/core/Box";
 
 const { REACT_APP_API_ENDPOINT } = process.env;
